Refactor GetPostDetails test and add AddReplyTo method
This commit refactors the GetPostDetails test in the PostReaderTests file to use AutoData generation and removes commented out code. Furthermore, an AddReplyTo method is added in the Post model to encapsulate the logic of assigning the ReplyTo property. This enhances the readability of the test and neatness of the Post model.