Add message serialization tests, update message interface, add new serializer
A new SystemTextMessageSerializer class an accompanying test class were added to verify the proper serialization and deserialization of messages. The IMessage interface was updated to allow setting of the MessageId and CreatedAt properties. A UserIdEnricher class was modified for sealing. A dependency on AutoFixture.Xunit2 was added to the package references. These changes were necessary to provide robust serialization functionality and allow testing of serialized messages.