Refactor MessageName to use full class name, remove Name property
Refactored MessageName in RabbitMqPublisherConfig and RabbtMqSubscriptionConfig to use the full class name of the message type instead of just the type name to uniquely identify message types. Removed the Name property from the IMessage interface and its implementations as it was redundant and could potentially cause conflicts due to non-unique names. This change ensures a unique and consistent identifier for each message type in the context of messaging across services.